[wasm2js] Fix base64 encoding (#1670)
The static std::string base64Encode(std::vector<char> &data) { uses signed char in input data. The ((int)data[0]) converts it the signed int, making '\xFF' char into -1. The patch fixes casting.

+++ src/wasm2js.h
@@ -447,10 +447,10 @@ static std::string base64Encode(std::vector<char> &data) {
     "0123456789+/";
 
   while (i + 3 <= data.size()) {
-    int bits =
-      (((int) data[i + 0]) << 16) |
-      (((int) data[i + 1]) << 8) |
-      (((int) data[i + 2]) << 0);
+    uint32_t bits =
+      (((uint32_t)(uint8_t) data[i + 0]) << 16) |
+      (((uint32_t)(uint8_t) data[i + 1]) << 8) |
+      (((uint32_t)(uint8_t) data[i + 2]) << 0);
     ret += alphabet[(bits >> 18) & 0x3f];
     ret += alphabet[(bits >> 12) & 0x3f];
     ret += alphabet[(bits >> 6) & 0x3f];
@@ -459,15 +459,15 @@ static std::string base64Encode(std::vector<char> &data) {
   }
 
   if (i + 2 == data.size()) {
-    int bits =
-      (((int) data[i + 0]) << 8) |
-      (((int) data[i + 1]) << 0);
+    uint32_t bits =
+      (((uint32_t)(uint8_t) data[i + 0]) << 8) |
+      (((uint32_t)(uint8_t) data[i + 1]) << 0);
     ret += alphabet[(bits >> 10) & 0x3f];
     ret += alphabet[(bits >> 4) & 0x3f];
     ret += alphabet[(bits << 2) & 0x3f];
     ret += '=';
   } else if (i + 1 == data.size()) {
-    int bits =(int) data[i + 0];
+    uint32_t bits = (uint32_t)(uint8_t) data[i + 0];
     ret += alphabet[(bits >> 2) & 0x3f];
     ret += alphabet[(bits << 4) & 0x3f];
     ret += '=';
